requirements.txt
discord.py>=2.3.0
cc-ampapi>=1.0.0